# Conflicts:
#	bunhobono/backend/src/main/resources/sql/bono_db_기존DB_업데이트 한 후 이거 실행 하면 member 포함 됨.sql
#	bunhobono/backend/src/main/resources/sql/bono_demo_reset_member_v2.sql

+++ bunhobono/backend/src/main/java/api/trash_p/TrashDTO.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 @Data
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class TrashDTO {
-    private Long trashNo;           // 휴지통 고유 번호
+    private Integer trashNo;        // 휴지통 고유 번호
     private String dataType;        // CAMERA_DATA, CAR_LOG, NOTICE
     private Integer originalNo;     // 원본 테이블의 PK 번호
     private String carNo;            // 삭제 데이터의 차량번호
